requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Click==7.0
coincurve==9.0.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cytoolz==0.9.0.1
defer==1.0.6
distro-info==0.18
eth-hash==0.2.0
eth-typing==1.3.0
eth-utils==1.2.2
ethereum==2.3.2
ethereum-serpent==2.0.7
Flask==1.0.2
future==0.17.1
httplib2==0.9.2
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
oauth==1.0.1
olefile==0.45.1
pbkdf2==1.3
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.7
py-ecc==1.4.3
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.0
pycups==1.9.73
pyethash==0.1.27
pygame==1.9.4
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pysha3==1.0.2
python-apt==1.6.3
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
repoze.lru==0.7
requests==2.18.4
requests-unixsocket==0.1.5
rlp==1.0.3
scrypt==0.8.6
SecretStorage==2.3.1
simpleaudio==1.0.2
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
toolz==0.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==0.14.1
xkit==0.0.0
zope.interface==4.3.2
